Initial commit
This is the first rev of mysql-sniffer, rewritten from Perl to Go, using the gopcap module from Andreas Krennmair.

This program is designed to sniff MySQL packets, pull out queries, and summarize them in a nice way for you to view and get an overview of what your database is doing while it's hammering away at some traffic.